Added Specular Material
-Added a ‘specular’ material in the materials folder that mimics the
appearance of the p3d.in setup I showed you guys. I made it a bit
lighter since its very dark as is, and we don’t want it to blend in
with the black of space. I imagine we’ll want to tweak it a bit. Lights
will help too when we get to those