Added Animations
-Animations are in place for each module (with the exception of the
default body as it has none). Each module (and the level) has an
animation controller (though nothing is hooked up ATM), as well as idle
and action animations to be utilized via code.

-The level’s middle pillar(s) is animated to slowly rotate. I opted to
not make the cube towers piece apart and together; it looked too busy
and I felt that it’d be more confusing (in terms of navigation /
landmarks) that cool

-Each module has a sub-mesh, allowing different materials for the body
and lighted areas

*****NOTE*****
When exporting as an fix with animations, everything gets scaled
differently, so we’ll likely need to resize the ships/modules. Scale
factor may be 100x what it was previously (that was the case for the
level’s model)